+++ hudi-common/src/test/java/org/apache/hudi/common/fs/TestStorageSchemes.java
@@ -19,7 +19,7 @@
 package org.apache.hudi.common.fs;
 
 import org.apache.hudi.hadoop.fs.HoodieWrapperFileSystem;
-import org.apache.hudi.io.storage.StorageSchemes;
+import org.apache.hudi.storage.StorageSchemes;
 
 import org.apache.hadoop.fs.Path;
 import org.junit.jupiter.api.Test;
