+++ hudi-io/src/main/java/org/apache/hudi/storage/StorageSchemes.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hudi.io.storage;
+package org.apache.hudi.storage;
 
 import java.util.Arrays;
